updated for Rails 4+ compatibility

+++ db/migrate/025_fractional_points.rb
@@ -1,12 +1,16 @@
-class FractionalPoints < ActiveRecord::Migration
+class FractionalPoints < ActiveRecord::Migration[4.2]
   def self.up
-    add_column :issues, :fractional_story_points, :float
-    execute "update issues set fractional_story_points = story_points"
-    remove_column :issues, :story_points
+    if column_exists? :issues, :story_points
+      add_column :issues, :fractional_story_points, :float
+      execute "update issues set fractional_story_points = story_points"
+  
+      change_column :issues, :story_points, :float
+      execute "update issues set story_points = fractional_story_points"
+      remove_column :issues, :fractional_story_points
+    else
+      add_column :issues, :story_points, :float
+    end
 
-    add_column :issues, :story_points, :float
-    execute "update issues set story_points = fractional_story_points"
-    remove_column :issues, :fractional_story_points
   end
 
   def self.down
